+++ pyhomebroker/__init__.py
@@ -19,7 +19,7 @@
 # limitations under the License.
 #
 
-__version__ = '0.4'
+__version__ = '0.41'
 __author__ = 'Diego Degese'
 
 from .home_broker import HomeBroker